requirements.txt
annotated-types
certifi
charset-normalizer
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
python-dotenv
requests
rsa
ruff
tqdm
typing-inspection
typing_extensions
uritemplate
urllib3
